backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
motor>=3.3.0
pydantic>=2.5.0
python-multipart>=0.0.6
websockets>=12.0
redis>=5.0.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0
pymongo>=4.6.0